Intelligent-Speech-Therapy-Platform-with-Adaptive-Exercises-and-Progress-Tracking_Feb_Batch-8_2026/backend/requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-dotenv==1.2.1
pydantic[email]==2.12.5
email-validator==2.3.0
